requirements.txt
telethon>=1.28.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
cryptg>=0.4.0  # Optional, for better performance
pillow>=10.0.0  # For handling media in Telegram
aiohttp>=3.8.0  # Required by Telethon for making API calls

